* master: (67 commits)
  sc_ver.h: bump version date
  examples/build-msvc: re-add dropped Makefile
  examples/build-unix: re-add dropped Makefile
  Fix MSBUILD-based batch example execution
  TLM examples: Update MSVC command-line build setup
  vsvars.bat: sync with SystemC regressions repository
  Cleanup MSVC project files for SystemC/TLM examples
  Remove MSVC warnings in (or triggered from) SystemC/TLM examples
  Update GNU make setup for SystemC & TLM examples
  Automake: update examples to include updated MSVC setup
  sc_vector: add missing SC_API annotation
  MSVC: Avoid warning C4267 (model side)
  VCD/WIF: hide MSVC warning
  sc_signal: cleanups for MSVC (DLL) build
  sc_ver.h: bump version date
  MSVC: minor update to MSVC'2010 project files
  Simplify explicit template instantiations
  Drop explicit std::vector template instantiations
  sc_ver.h: bump version date
  Keep sc_(z)string_view implementation-defined
  ...

+++ examples/sysc/fft/fft_fxpt/fft.cpp
@@ -249,7 +249,7 @@ void fft::entry()
              real[index] = tmp_real;
              imag[index] = tmp_imag;
 	    
-	     i = i + (len << 1);   
+             i = (short)(i + (len << 1));
           }
 
         //Remaining Iterations: Use Stored W
@@ -281,7 +281,7 @@ void fft::entry()
               real[index2] = real2_out;
               imag[index2] = imag2_out; 
 
-              i = i + (len << 1);
+              i = (short)(i + (len << 1));
             }
             windex = windex + incr;
             j++;
@@ -310,7 +310,7 @@ void fft::entry()
        bits_index[2]= bits_i[1];
        bits_index[1]= bits_i[2];
        bits_index[0]= bits_i[3];
-       index = bits_index;
+       index = (short)bits_index;
        real1 = real[index];
        imag1 = imag[index];
        out_real.write(real1);
